requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
faiss-cpu
openai
python-dotenv
apscheduler
pydantic
httpx
